requirements-edge.txt
opencv-python-headless
depthai
google-genai
blobconverter
tenacity
pydantic-settings>=2.0
